# target/classes/api.yaml
openapi: 3.0.3
info:
  title: ms_mandarina_v1
  description: Servicio encargado de persistir las mandarinas
  version: 1.0.0

paths:
  /mandarina:
    post:
      tags:
        - Creación de una mandarina
      operationId: createMandarina
      summary: Crear una nueva mandarina
      description: Este servicio crea una nueva mandarina en base de datos con los parámetros recibidos en el servicio
      requestBody:
        description: Datos de entrada
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mandarina'
        required: true
      responses:
        201:
          $ref: '#/components/responses/mandarina_rs'
        400:
          $ref: '#/components/responses/bad_request'
        500:
          $ref: '#/components/responses/internal_error'

components:
  responses:
    mandarina_rs:
      description: En caso de que la mandarina se haya creado correctamente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/mandarina'
    bad_request:
      description: En caso de que no se envien los datos necesarios para la creación de la mandarina o los mismos tengan un formato erroneo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error_response'
    internal_error:
      description: En caso de error general en el servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error_response'

  schemas:
    mandarina:
      required:
        - name
      type: object
      properties:
        id:
          type: string
          description: Identificador de la mandarina
          readOnly: true
        name:
          type: string
          description: Nombre de la mandarina
    error_response:
      type: object
      properties:
        error:
          required:
            - message
            - request_id
          type: object
          description: Payload estandard
          properties:
            message:
              type: string
              description: Describe la causa del error. Las descripciones se deberán
                componer de textos fijos para cada tipo de error producido, por lo
                que deben ser generadas a partir de una estructura de datos como un
                enumerado.
            request_id:
              type: string
              description: Describe un código de petición único. Esto permite disponer
                de trazabilidad en caso de necesitar inspeccionar logs.
            details:
              type: string
              description: Contiene un objeto con el detalle adicional de errores
                que desee aportar cada product owner del aplicativo sobre el que se
                desarrolla el servicio REST y que debe ser definido por este. Puede
                contener arrays con mayor detalle.
            code:
              type: integer
              description: Respuesta proporcionada por las aplicaciones en caso de
                error. Dicho valor será utilizado como Código de Error de Aplicación,
                siendo su empleo opcional y de libre uso por parte de los responsables
                de las aplicaciones para gestionar sus propios códigos de retorno
                ante errores, aunque en caso de utilizarse, es obligatorio documentar
                apropiadamente el significado de cada uno de los posibles valores
                devueltos en dicho atributo, pudiendo documentarse en el propio fichero
                OpenAPI o pudiendo ser documentado apropiadamente (subiendo documentación
                adicional al API con los posibles códigos de error o un enlace a dicha
                documentación a un enlace externo) para que sea fácilmente accesible
                en caso de precisarse consultar.